fix : aop 수정

+++ src/main/java/store/buzzbook/front/common/aop/UserJwtAop.java
@@ -36,7 +36,7 @@ public class UserJwtAop {
 		//회원
 		Map<String, Object> claims = jwtService.getInfoMapFromJwt(request);
 
-		Long userId = (Long)claims.get(JwtService.USER_ID);
+		Long userId = ((Integer)claims.get(JwtService.USER_ID)).longValue();
 		String loginId = (String)claims.get(JwtService.LOGIN_ID);
 		String role = (String)claims.get(JwtService.ROLE);
